Refactor docs handler and config into dedicated packages
Further preparation for SES and updater packages.
